Fix all clippy warnings in firmware crate
- Add Default impls for DebounceState, EdgeDetector, CvChannel,
  CvInputs, Framebuffer, QuadratureState, ButtonDebounce, MidiParser
- Use abs_diff() instead of manual if/else in CvChannel::changed
- Remove unnecessary u64 cast in cv_output
- Fix overlapping match arms in MIDI parser (0xF8..=0xFF vs CLOCK)
- Use enumerate() instead of range indexing in LED track colors
- Collapse nested if in periodic save
- Allow too_many_arguments on process_tick
- Use is_multiple_of() for clock output step boundary check
- Allow dead_code on EdgeDetector (used in host testing only)

+++ crates/firmware/src/cv_output.rs
@@ -99,7 +99,7 @@ impl CvOutputProcessor {
         let ch = &mut self.channels[output];
 
         // Calculate samples in this step window
-        ch.step_samples = (step_duration_us as u64 * self.render_rate as u64 / 1_000_000) as u32;
+        ch.step_samples = (step_duration_us * self.render_rate as u64 / 1_000_000) as u32;
         if ch.step_samples == 0 {
             ch.step_samples = 1;
         }
